+++ bonus/src/floors_ceilings_bonus.c
@@ -1,16 +1,16 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   floors_ceilings.c                                  :+:      :+:    :+:   */
+/*   floors_ceilings_bonus.c                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: mraymond <mraymond@student.42quebec.com    +#+  +:+       +#+        */
+/*   By: anhebert <anhebert@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/30 14:15:38 by mraymond          #+#    #+#             */
-/*   Updated: 2022/11/30 14:43:17 by mraymond         ###   ########.fr       */
+/*   Updated: 2022/12/01 11:29:59 by anhebert         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../include/cub3d.h"
+#include "../include/cub3d_bonus.h"
 
 bool	is_seperator(char c, bool *is_virg)
 {
