+++ bonus/src/mlx_stuff_bonus.c
@@ -1,16 +1,16 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   mlx_stuff.c                                        :+:      :+:    :+:   */
+/*   mlx_stuff_bonus.c                                  :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: mraymond <mraymond@student.42quebec.com    +#+  +:+       +#+        */
+/*   By: anhebert <anhebert@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/30 14:16:13 by mraymond          #+#    #+#             */
-/*   Updated: 2022/11/30 14:16:14 by mraymond         ###   ########.fr       */
+/*   Updated: 2022/12/01 12:16:50 by anhebert         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../include/cub3d.h"
+#include "../include/cub3d_bonus.h"
 
 void	vars_mlx_init(t_vars *vars)
 {
@@ -35,6 +35,9 @@ int	render_next_frame(t_vars *vars)
 			&vars->mlx.line_len, &vars->mlx.endian);
 	mlx_hook(vars->mlx.win, 2, 0, keypress_handler, vars);
 	raycast_main_loop(vars);
+	//minimap_init(vars);
+	draw_map(vars);
+	draw_perso(vars);
 	mlx_put_image_to_window(vars->mlx.mlx, vars->mlx.win,
 		vars->mlx.img[i[0]], 0, 0);
 	mlx_destroy_image (vars->mlx.mlx, vars->mlx.img[i[1]]);
