+++ bonus/src/variables_identification_tools_bonus.c
@@ -1,16 +1,16 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   variables_identification_tools.c                   :+:      :+:    :+:   */
+/*   variables_identification_tools_bonus.c             :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: mraymond <mraymond@student.42quebec.com    +#+  +:+       +#+        */
+/*   By: anhebert <anhebert@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/30 14:16:53 by mraymond          #+#    #+#             */
-/*   Updated: 2022/11/30 14:16:54 by mraymond         ###   ########.fr       */
+/*   Updated: 2022/12/01 11:30:36 by anhebert         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../include/cub3d.h"
+#include "../include/cub3d_bonus.h"
 
 bool	is_mapchar(char c)
 {
