+++ bonus/src/various_tools_bonus.c
@@ -1,16 +1,16 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   various_tools.c                                    :+:      :+:    :+:   */
+/*   various_tools_bonus.c                              :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: mraymond <mraymond@student.42quebec.com    +#+  +:+       +#+        */
+/*   By: anhebert <anhebert@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/30 14:16:57 by mraymond          #+#    #+#             */
-/*   Updated: 2022/11/30 14:16:58 by mraymond         ###   ########.fr       */
+/*   Updated: 2022/12/01 11:30:38 by anhebert         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../include/cub3d.h"
+#include "../include/cub3d_bonus.h"
 
 char	*ft_strcat(char *s1, int len)
 {
